Manual merge of v0.9a:graphs and tensorrt:graphs
